+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,6 +1,5 @@
-package kitchenpos.domain.product;
+package kitchenpos.product.domain;
 
-import kitchenpos.product.domain.Product;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
