+++ src/test/java/kitchenpos/table/application/TableServiceTest.java
@@ -1,7 +1,5 @@
-package kitchenpos.application;
+package kitchenpos.table.application;
 
-import kitchenpos.table.application.TableService;
-import kitchenpos.table.application.TableValidator;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.TableGroup;
@@ -25,7 +23,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
 
 @DisplayName("주문 테이블 관련 테스트")
